Change prompter class to LintPrivate & move helper_methods to that class and update bracket_checker method & fix rubocop errors
