--- lib/prompter.rb
@@ -1,5 +1,0 @@
-require_relative '../lib/linters.rb'
-require 'colorize'
-class Prompter
-end
-
