+++ test_files/lint-bracket-missing.css
@@ -1,4 +1,4 @@
-.some-title {
+.some-title 
   font-weight: lighter;
   font-weight: lighter;
 }
@@ -7,11 +7,11 @@
   font-weight: lighter;
 }
 
-.some-title {
+.some-title 
   font-weight: lighter;
   font-weight: lighter;
 
 
-.some-title {
+.some-title 
   font-weight: lighter;
-}
+
